+++ apps/doc/src/routes/(rails)/+layout.svelte
@@ -53,8 +53,14 @@
 		},
 		{
 			zone: zoneContent,
-			hug: layout === 'steam' || layout === 'tram' || layout === 'urbanist',
+			hug:
+				layout === 'steam' ||
+				layout === 'tram' ||
+				layout === 'urbanist' ||
+				layout === 'voyager',
 			grid: true,
+			container: 'text',
+			containerSize: '2xl',
 		},
 		{
 			zone: zoneFooter,