+++ packages/ui/src/lib/types/style.ts
@@ -1,4 +1,10 @@
-export type UiContainer = 'burrito' | 'taco' | 'frame' | 'ravioli' | 'raviolink'
+export type UiContainer =
+	| 'burrito'
+	| 'taco'
+	| 'frame'
+	| 'ravioli'
+	| 'raviolink'
+	| 'text'
 
 export type UiLayout =
 	| 'stack'